Fix bug in music pedagogy page.
Exhibits expects that the body of a WebAnnotation to be an array,
but it should be a JSON object.  This is a temporary fix by pointing
at a static manifest that changes the data type. This should have
been done previously, but I guess it was missed. We never noticed
because the manifest was cached.